Fixed disabled warning violations.
