+++ Common/.editorconfig
@@ -144,6 +144,5 @@ csharp_space_between_square_brackets = false
 csharp_preserve_single_line_blocks = true
 csharp_preserve_single_line_statements = true
 
-
 # CA1307: Specify StringComparison
 dotnet_diagnostic.CA1307.severity = warning
