+++ PluginFramework/.editorconfig
@@ -144,9 +144,5 @@ csharp_space_between_square_brackets = false
 csharp_preserve_single_line_blocks = true
 csharp_preserve_single_line_statements = true
 
-
 # S101: Types should be named in PascalCase
 dotnet_diagnostic.S101.severity = none
-
-# S4220: Events should have proper arguments
-dotnet_diagnostic.S4220.severity = none
