+++ ZipAssembly/.editorconfig
@@ -144,7 +144,6 @@ csharp_space_between_square_brackets = false
 csharp_preserve_single_line_blocks = true
 csharp_preserve_single_line_statements = true
 
-
 # S1905: Redundant casts should not be used
 dotnet_diagnostic.S1905.severity = none
